Estoque-API/requirements/base.txt
python-slugify==8.0.4  # https://github.com/un33k/python-slugify
Pillow==10.4.0  # https://github.com/python-pillow/Pillow
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.7.0  # https://github.com/evansd/whitenoise
redis==5.0.7  # https://github.com/redis/redis-py
hiredis==2.3.2  # https://github.com/redis/hiredis-py
celery==5.4.0  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.6.0  # https://github.com/celery/django-celery-beat
flower==2.0.1  # https://github.com/mher/flower
uvicorn[standard]==0.30.1  # https://github.com/encode/uvicorn
uvicorn-worker==0.2.0  # https://github.com/Kludex/uvicorn-worker

# Django
# ------------------------------------------------------------------------------
django==4.2.13  # pyup: < 5.0  # https://www.djangoproject.com/
django-environ==0.11.2  # https://github.com/joke2k/django-environ
django-unfold==0.30.0
django-filter==24.2
django-simple-history==3.7.0
django-import-export==4.1.0
django-guardian==2.4.0
django-model-utils==4.5.1  # https://github.com/jazzband/django-model-utils
django-allauth[mfa]==0.63.3  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.2  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==2024.2  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.4.0  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.15.2  # https://github.com/encode/django-rest-framework
django-cors-headers==4.4.0  # https://github.com/adamchainz/django-cors-headers
# DRF-spectacular for api documentation
drf-spectacular==0.27.2  # https://github.com/tfranzel/drf-spectacular
django-webpack-loader==3.1.0  # https://github.com/django-webpack/django-webpack-loader

django-extensions==3.2.3  # https://github.com/django-extensions/django-extensions